chore: Remove useless play 25 directive, proper project naming, drop empty play-2.5
