--- dd-java-agent/instrumentation/play/play-2.5/build.gradle
@@ -1,30 +1,0 @@
-muzzle {
-  extraRepository("typesafe", "https://repo.typesafe.com/typesafe/maven-releases/")
-
-  pass {
-    name = "play25"
-    group = 'com.typesafe.play'
-    module = 'play_2.11'
-    versions = '[2.5.0,2.6)'
-    assertInverse = true
-  }
-}
-
-apply from: "$rootDir/gradle/java.gradle"
-
-testJvmConstraints {
-  // Play doesn't work with Java 9+ until 2.6.12
-  maxJavaVersion = JavaVersion.VERSION_1_8
-}
-
-repositories {
-  maven {
-    name = 'typesafe'
-    url = 'https://repo.typesafe.com/typesafe/maven-releases/'
-  }
-}
-
-dependencies {
-  implementation project(':dd-java-agent:instrumentation:play:play-2.4')
-  compileOnly group: 'com.typesafe.play', name: 'play_2.11', version: '2.5.0'
-}
