+++ dd-java-agent/instrumentation/play/play-appsec-2.5/src/main/java/datadog/trace/instrumentation/play25/appsec/FormUrlEncodedInstrumentation.java
@@ -1,24 +1,21 @@
 package datadog.trace.instrumentation.play25.appsec;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.returns;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
 import datadog.trace.agent.tooling.muzzle.Reference;
+import java.util.Map;
 
 @AutoService(InstrumenterModule.class)
-public class StatusHeaderInstrumentation extends InstrumenterModule.AppSec
+public class FormUrlEncodedInstrumentation extends InstrumenterModule.AppSec
     implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
-
-  public StatusHeaderInstrumentation() {
+  public FormUrlEncodedInstrumentation() {
     super("play");
-  }
-
-  @Override
-  public String muzzleDirective() {
-    return "play25only";
   }
 
   @Override
@@ -28,13 +25,24 @@ public class StatusHeaderInstrumentation extends InstrumenterModule.AppSec
 
   @Override
   public String instrumentedType() {
-    return "play.mvc.StatusHeader";
+    return "play.mvc.BodyParser$FormUrlEncoded";
+  }
+
+  @Override
+  public String[] helperClassNames() {
+    return new String[] {
+      packageName + ".BodyParserHelpers",
+    };
   }
 
   @Override
   public void methodAdvice(MethodTransformer transformer) {
     transformer.applyAdvice(
-        named("sendJson").and(takesArgument(0, named("com.fasterxml.jackson.databind.JsonNode"))),
-        packageName + ".StatusHeaderSendJsonAdvice");
+        named("parse")
+            .and(takesArguments(2))
+            .and(takesArgument(0, named("play.mvc.Http$RequestHeader")))
+            .and(takesArgument(1, named("akka.util.ByteString")))
+            .and(returns(Map.class)),
+        packageName + ".BodyParserFormUrlEncodedParseAdvice");
   }
 }
