+++ dd-java-agent/instrumentation/play/play-appsec-2.5/src/main/java/datadog/trace/instrumentation/play25/appsec/HttpErrorHandlerInstrumentation.java
@@ -32,11 +32,6 @@ public class HttpErrorHandlerInstrumentation extends InstrumenterModule.AppSec
   }
 
   @Override
-  public String muzzleDirective() {
-    return "play25only";
-  }
-
-  @Override
   public Reference[] additionalMuzzleReferences() {
     return MuzzleReferences.PLAY_25_ONLY;
   }
