+++ dd-java-agent/instrumentation/play/play-appsec-2.5/src/main/java/datadog/trace/instrumentation/play25/appsec/ResultsStatusInstrumentation.java
@@ -16,11 +16,6 @@ public class ResultsStatusInstrumentation extends InstrumenterModule.AppSec
   }
 
   @Override
-  public String muzzleDirective() {
-    return "play25only";
-  }
-
-  @Override
   public Reference[] additionalMuzzleReferences() {
     return MuzzleReferences.PLAY_25_ONLY;
   }
