+++ dd-java-agent/instrumentation/play/play-appsec-2.5/src/main/java/datadog/trace/instrumentation/play25/appsec/RoutingDslInstrumentation.java
@@ -20,11 +20,6 @@ public class RoutingDslInstrumentation extends InstrumenterModule.AppSec
   }
 
   @Override
-  public String muzzleDirective() {
-    return "play25only";
-  }
-
-  @Override
   public Reference[] additionalMuzzleReferences() {
     return MuzzleReferences.PLAY_25_ONLY;
   }
