+++ dd-java-agent/instrumentation/play/play-appsec-2.5/src/main/java/datadog/trace/instrumentation/play25/appsec/SirdPathExtractorInstrumentation.java
@@ -1,8 +1,7 @@
 package datadog.trace.instrumentation.play25.appsec;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.isStatic;
-import static net.bytebuddy.matcher.ElementMatchers.not;
+import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.namedOneOf;
 import static net.bytebuddy.matcher.ElementMatchers.returns;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
@@ -13,18 +12,13 @@ import datadog.trace.agent.tooling.InstrumenterModule;
 import datadog.trace.agent.tooling.muzzle.Reference;
 
 /**
- * @see play.core.routing.PathPattern#apply(String)
+ * @see play.api.routing.sird.PathExtractor
  */
 @AutoService(InstrumenterModule.class)
-public class PathPatternInstrumentation extends InstrumenterModule.AppSec
+public class SirdPathExtractorInstrumentation extends InstrumenterModule.AppSec
     implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
-  public PathPatternInstrumentation() {
+  public SirdPathExtractorInstrumentation() {
     super("play");
-  }
-
-  @Override
-  public String muzzleDirective() {
-    return "play25only";
   }
 
   @Override
@@ -33,25 +27,24 @@ public class PathPatternInstrumentation extends InstrumenterModule.AppSec
   }
 
   @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      packageName + ".PathExtractionHelpers",
-    };
-  }
-
-  @Override
   public String instrumentedType() {
-    return "play.core.routing.PathPattern";
+    return "play.api.routing.sird.PathExtractor";
   }
 
   @Override
   public void methodAdvice(MethodTransformer transformer) {
     transformer.applyAdvice(
-        named("apply")
-            .and(not(isStatic()))
+        namedOneOf("extract", "play$api$routing$sird$PathExtractor$$extract")
             .and(takesArguments(1))
             .and(takesArgument(0, String.class))
             .and(returns(named("scala.Option"))),
-        packageName + ".PathPatternApplyAdvice");
+        packageName + ".SirdPathExtractorExtractAdvice");
+  }
+
+  @Override
+  public String[] helperClassNames() {
+    return new String[] {
+      packageName + ".PathExtractionHelpers",
+    };
   }
 }
