+++ settings.gradle.kts
@@ -498,8 +498,7 @@ include(
   ":dd-java-agent:instrumentation:play-ws:play-ws-common",
   ":dd-java-agent:instrumentation:play:play-2.3",
   ":dd-java-agent:instrumentation:play:play-2.4",
-  ":dd-java-agent:instrumentation:play:play-2.5",
-  ":dd-java-agent:instrumentation:play:play-2.5-appsec",
+  ":dd-java-agent:instrumentation:play:play-appsec-2.5",
   ":dd-java-agent:instrumentation:play:play-2.6",
   ":dd-java-agent:instrumentation:protobuf-3.0",
   ":dd-java-agent:instrumentation:quartz-2.0",
